Mega cart update: fixed issue that wasn't allowing the page to re-render the content if the cart was empty and added the posibility to increase and decrease the amount of products you want to buy.

+++ src/data/useCartStore.js
@@ -1,17 +1,35 @@
 import { create } from 'zustand';
 
+// Inicializamos el carrito creando uno.
 export const useCartStore = create((set) => ({
+  // Se crea el array 'Cart', nuestro carrito
   cart: JSON.parse(typeof window !== "undefined" && localStorage.getItem('cart')) || [],
-  
-  //Método para añadir al carrito y manejar los estados globales
+
+  //A continuación, los métodos que podremos usar para manipular el carrito:
+
+  // Primer método: añadir items al carrito. Si el producto existe, se devuelve con normalidad. 
   addToCart: (idProducto) => set((state) => {
-    const updatedCart = [...state.cart, idProducto];
+    const existingProductIndex = state.cart.findIndex(item => item.id === idProducto);
+
+    if (existingProductIndex >= 0) {
+      return state; 
+    }
+
+    const updatedCart = [...state.cart, { id: idProducto, cantidad: 1 }];
     localStorage.setItem('cart', JSON.stringify(updatedCart));
     return { cart: updatedCart };
   }),
 
   removeFromCart: (idProducto) => set((state) => {
-    const updatedCart = state.cart.filter(id => id !== idProducto);
+    const updatedCart = state.cart.filter(item => item.id !== idProducto);
+    localStorage.setItem('cart', JSON.stringify(updatedCart));
+    return { cart: updatedCart };
+  }),
+
+  updateCartItem: (idProducto, cantidad) => set((state) => {
+    const updatedCart = state.cart.map(item => 
+      item.id === idProducto ? { ...item, cantidad } : item
+    );
     localStorage.setItem('cart', JSON.stringify(updatedCart));
     return { cart: updatedCart };
   }),
